Adding preivew component for homepage - Load more link will go to full container
