+++ src/containers/Release/index.js
@@ -2,26 +2,41 @@ import React, { Component } from 'react';
 
 
 class Release extends Component {
-    
     state = {
-        showInfo: false,
+        showInfoState: false
+    }
+
+    clickHandler = (event, item) => {
+        this.setState({
+            showInfo: !this.state.showInfo
+        })
     }
 
     render() {
+
         return (
-            <div className="release"
-                id={release.releaseNumber}
-                onClick={release.onClick}
-            >
-                <img src={release.img} className="cover" alt={release.name} />
-                <div className="description">
-                    <div>{release.name}</div>
-                    <div>{release.artist}</div>
+            (this.state.showInfo) ?
+                <div className="release"
+                    id={this.props.data.releaseNumber}
+                    onClick={this.clickHandler}
+                >
+                    <div>{this.props.data.name}</div>
+                    <div>{this.props.data.artist}</div>
+                    <div> Release: {this.props.data.releaseNumber}</div>
+
                 </div>
-            </div>
+                :
+                <div className="release"
+                    id={this.props.data.releaseNumber}
+                    onClick={this.clickHandler}
+                >
+                    <img src={this.props.data.img}
+                        className="cover"
+                        alt={this.props.data.name}
+                    />
+                </div>
         )
     }
-   
 }
 
-export default Release
+export default Release;
