+++ src/containers/Releases/index.js
@@ -1,69 +1,31 @@
 import React, { Component } from 'react';
+import { withRouter } from 'react-router-dom';
 import releaseData from './releaseData.js';
-
-
-class Release extends Component {
-    state = {
-        showInfoState: false
-    }
-
-    clickHandler = (event, item) => {
-        this.setState({
-            showInfo: !this.state.showInfo
-        })
-    }
-
-    render() {
-
-        return (
-            (this.state.showInfo) ?
-                <div className="release"
-                    id={this.props.data.releaseNumber}
-                    onClick={this.clickHandler}
-                >
-                    <div>{this.props.data.name}</div>
-                    <div>{this.props.data.artist}</div>
-                    <div> Release: {this.props.data.releaseNumber}</div>
-
-                </div>
-                :
-                <div className="release"
-                    id={this.props.data.releaseNumber}
-                    onClick={this.clickHandler}
-                >
-                    <img src={this.props.data.img} 
-                        className="cover" 
-                        alt={this.props.data.name}
-                    />
-                </div>
-        )
-    }
-
-}
-
+import Release from '../Release/index.js';
 
 
 class Releases extends Component {
 
     state = {
-        itemsToRender: 2,
+        itemsToRender: (releaseData.length-1),
     }
 
 
     loadMoreHandler = () => {
-        this.setState({
-            itemsToRender: (this.state.itemsToRender + 4)
-        })
+        this.props.history.push('/releases');
+
     }
 
+
     render() {
-        console.log('state check ' + this.state.showInfoState)
+        console.log('state check ' + this.state.itemsToRender)
+
         return (
             <div className="content" id="releases">
                 <div className="sectionHeader">Releases</div>
                 <div className="displayGrid">
                     {
-                        releaseData.slice(0, this.state.itemsToRender).map(
+                        releaseData.map(
                             (release, index) => {
                                 return (
                                     <Release
@@ -76,20 +38,9 @@ class Releases extends Component {
                         )
                     }
                 </div>
-                {
-                    (releaseData.length > this.state.itemsToRender) ?
-                        <button className="loadMore" 
-                            id="loadReleases" 
-                            onClick={this.loadMoreHandler}
-                        >
-                            Load more...
-                        </button>
-                        : null
-                }
-
             </div>
         )
     }
 }
 
-export default Releases;
+export default withRouter(Releases);
