+++ src/containers/Sidebar/index.css
@@ -23,4 +23,5 @@
   height: 17vmin;
   width: 15vmin;
   color: black;
+  margin-right: 5vmin;
 }
